# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YPuddin Train Studio",
    "version": "0.5.0"
  },
  "paths": {
...
    "/api/projects/{pid}/versions/{vid}/pipeline": {
      "get": {
        "summary": "Get Pipeline",
        "operationId": "get_pipeline_api_projects__pid__versions__vid__pipeline_get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "vid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{pid}/versions/{vid}/pipeline/operations": {
      "post": {
        "summary": "Start Operation",
        "operationId": "start_operation_api_projects__pid__versions__vid__pipeline_operations_post",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "vid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineOperation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dataset-pipeline/operations/{oid}": {
      "get": {
        "summary": "Get Operation",
        "operationId": "get_operation_api_dataset_pipeline_operations__oid__get",
        "parameters": [
          {
            "name": "oid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Oid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineOperation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dataset-pipeline/operations/{oid}/cancel": {
      "post": {
        "summary": "Cancel Operation",
        "operationId": "cancel_operation_api_dataset_pipeline_operations__oid__cancel_post",
        "parameters": [
          {
            "name": "oid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Oid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineOperation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dataset-pipeline/operations/{oid}/retry": {
      "post": {
        "summary": "Retry Operation",
        "operationId": "retry_operation_api_dataset_pipeline_operations__oid__retry_post",
        "parameters": [
          {
            "name": "oid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Oid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineOperation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dataset-tagging/status": {
      "get": {
        "summary": "Tagging Status",
        "operationId": "tagging_status_api_dataset_tagging_status_get",
        "parameters": [
          {
            "name": "model_path",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Path"
            }
          },
          {
            "name": "tags_path",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tags Path"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "cpu",
                "cuda"
              ],
              "type": "string",
              "default": "cpu",
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaggingStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaptionOptions": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "fill_missing",
              "append",
              "remove",
              "replace"
            ],
            "title": "Mode",
            "default": "fill_missing"
          },
          "text": {
            "type": "string",
            "maxLength": 32000,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "CaptionOptions"
      },
...
      "CropRectangle": {
        "properties": {
          "x": {
            "type": "integer",
            "minimum": 0.0,
            "title": "X"
          },
          "y": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Y"
          },
          "width": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "title": "CropRectangle"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocalTaggingModel": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "model_path": {
            "type": "string",
            "title": "Model Path"
          },
          "tags_path": {
            "type": "string",
            "title": "Tags Path"
          }
        },
        "type": "object",
        "required": [
          "name",
          "path",
          "model_path",
          "tags_path"
        ],
        "title": "LocalTaggingModel"
      },
...
      "PipelineImage": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "rel_path": {
            "type": "string",
            "minLength": 1,
            "title": "Rel Path"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "rel_path"
        ],
        "title": "PipelineImage"
      },
      "PipelineOperation": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "version_id": {
            "type": "string",
            "title": "Version Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "done": {
            "type": "integer",
            "title": "Done"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "updated_at": {
            "type": "number",
            "title": "Updated At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "request": {
            "additionalProperties": true,
            "type": "object",
            "title": "Request"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "logs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Logs"
          },
          "can_undo": {
            "type": "boolean",
            "title": "Can Undo"
          },
          "can_cancel": {
            "type": "boolean",
            "title": "Can Cancel"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "version_id",
          "action",
          "status",
          "phase",
          "done",
          "total",
          "created_at",
          "updated_at",
          "request",
          "result",
          "logs",
          "can_undo",
          "can_cancel"
        ],
        "title": "PipelineOperation"
      },
      "PipelineRequest": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "inspect",
              "exclude",
              "restore",
              "preprocess",
              "captions",
              "tag",
              "prepare"
            ],
            "title": "Action"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/PipelineImage"
            },
            "type": "array",
            "maxItems": 20000,
            "title": "Images"
          },
          "restore_operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Restore Operation Id"
          },
          "preprocess": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreprocessOptions"
              },
              {
                "type": "null"
              }
            ]
          },
          "captions": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaptionOptions"
              },
              {
                "type": "null"
              }
            ]
          },
          "tagging": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaggingOptions"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "PipelineRequest"
      },
      "PipelineSnapshot": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "version_id": {
            "type": "string",
            "title": "Version Id"
          },
          "signature": {
            "type": "string",
            "title": "Signature"
          },
          "inspection": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspection"
          },
          "plan": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan"
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/PipelineOperation"
            },
            "type": "array",
            "title": "Operations"
          },
          "datasets": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Datasets"
          },
          "busy": {
            "type": "boolean",
            "title": "Busy"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived"
          },
          "ready_to_train": {
            "type": "boolean",
            "title": "Ready To Train"
          },
          "prepared_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prepared Job Id"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "version_id",
          "signature",
          "inspection",
          "plan",
          "operations",
          "datasets",
          "busy",
          "archived",
          "ready_to_train",
          "prepared_job_id",
          "stale"
        ],
        "title": "PipelineSnapshot"
      },
...
      "PreprocessOptions": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "resize",
              "center_crop",
              "crop_rect"
            ],
            "title": "Mode",
            "default": "resize"
          },
          "width": {
            "type": "integer",
            "maximum": 8192.0,
            "minimum": 16.0,
            "title": "Width",
            "default": 1024
          },
          "height": {
            "type": "integer",
            "maximum": 8192.0,
            "minimum": 16.0,
            "title": "Height",
            "default": 1024
          },
          "allow_upscale": {
            "type": "boolean",
            "title": "Allow Upscale",
            "default": false
          },
          "crop": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CropRectangle"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "PreprocessOptions"
      },
...
      "TaggingOptions": {
        "properties": {
          "model_path": {
            "type": "string",
            "minLength": 1,
            "title": "Model Path"
          },
          "tags_path": {
            "type": "string",
            "minLength": 1,
            "title": "Tags Path"
          },
          "general_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "General Threshold",
            "default": 0.35
          },
          "character_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Character Threshold",
            "default": 0.85
          },
          "provider": {
            "type": "string",
            "enum": [
              "cpu",
              "cuda"
            ],
            "title": "Provider",
            "default": "cpu"
          },
          "mode": {
            "type": "string",
            "enum": [
              "missing",
              "append",
              "overwrite"
            ],
            "title": "Mode",
            "default": "missing"
          },
          "trigger_word": {
            "type": "string",
            "maxLength": 1000,
            "title": "Trigger Word",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "model_path",
          "tags_path"
        ],
        "title": "TaggingOptions"
      },
      "TaggingStatus": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "runtime_available": {
            "type": "boolean",
            "title": "Runtime Available"
          },
          "runtime_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Version"
          },
          "runtime_providers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Runtime Providers"
          },
          "providers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Providers"
          },
          "provider": {
            "type": "string",
            "enum": [
              "cpu",
              "cuda"
            ],
            "title": "Provider"
          },
          "model_exists": {
            "type": "boolean",
            "title": "Model Exists"
          },
          "tags_exists": {
            "type": "boolean",
            "title": "Tags Exists"
          },
          "model_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Path"
          },
          "tags_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags Path"
          },
          "input_size": {
            "type": "integer",
            "title": "Input Size"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/LocalTaggingModel"
            },
            "type": "array",
            "title": "Models"
          },
          "recommended_model_dir": {
            "type": "string",
            "title": "Recommended Model Dir"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "available",
          "runtime_available",
          "runtime_version",
          "runtime_providers",
          "providers",
          "provider",
          "model_exists",
          "tags_exists",
          "model_path",
          "tags_path",
          "input_size",
          "models",
          "recommended_model_dir",
          "errors",
          "notes"
        ],
        "title": "TaggingStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}